+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Changelog
-### 🐞 Bug Fixes
-- Fixed a bug in multiplayer where lamps were synchronized immediately for all players. 
+### 🚀 Added
+- API to support multiple compatibility layers.